Refactored Database by seperating out collection types. npm package install save for cookieparser and session.
