+++ package.json
@@ -18,8 +18,11 @@
   },
   "dependencies": {
     "angular": "^1.6.3",
+    "bcrypt": "^1.0.2",
     "body-parser": "^1.17.1",
+    "cookie-parser": "^1.4.3",
     "express": "^4.15.0",
+    "express-session": "^1.15.1",
     "jquery": "^3.1.1",
     "mongoose": "^4.8.6",
     "mysql": "^2.13.0",
